Clean up code - see comments
