--- react-app/src/components/Brandon_NavBar/ProfileButton.js
@@ -1,1 +1,0 @@
-
