+++ react-app/src/components/BusinessDetails/BusinessDetails.css
@@ -109,6 +109,7 @@
 .Closedclosed-red {
     color: rgba(224, 7, 7, 1);
 }
+
 .Openclosed-red {
     color: #00C588;
     padding-right: 6px;
@@ -141,9 +142,10 @@
 .single-business-reviews-button {
     width: 100%;
 }
+
 .single-business-see-more-photos:hover {
     cursor: pointer;
-    background-color:rgba(254, 254, 254, 0.402);
+    background-color: rgba(254, 254, 254, 0.402);
     transition-property: background-color;
     transition-duration: 0.5s;
     transition-timing-function: ease-in;
@@ -172,6 +174,7 @@
     width: 80%;
     margin: auto;
 }
+
 #single-biz-stars {
     margin-bottom: 30px;
 }
@@ -196,11 +199,12 @@
     border-radius: 4px;
     font-size: 16px;
 }
+
 .single-business-review-bar-button:hover {
-background-color:rgb(255, 47, 47);
-transition-property: background-color;
-transition-duration: 0.5s;
-transition-timing-function: ease-in;
+    background-color: rgb(255, 47, 47);
+    transition-property: background-color;
+    transition-duration: 0.5s;
+    transition-timing-function: ease-in;
 }
 
 .single-business-location-hours {
@@ -357,8 +361,9 @@ transition-timing-function: ease-in;
 
 .photo-modal-close {
     display: flex;
-    justify-content:end;
+    justify-content: end;
 }
+
 .photo-modal-close:hover {
     cursor: pointer;
 }
