+++ react-app/src/components/CreateReviewImgForm/CreateReviewImgForm.css
@@ -40,6 +40,7 @@
     border: 1.3px solid #CCC;
     transition: 0.8s ease background-color;
 }
+
 #add-review-img-submit-button:hover {
     background-color: rgb(255, 17, 17);
     box-shadow: 0 0 2px 0 rgba(0, 0, 0, 0.5);
