+++ react-app/src/components/ErrorPage/ErrorPage.css
@@ -1,26 +1,26 @@
 #apologies {
-  color: #A9826E;
-  margin: 0;
-  font-weight: 400;
+    color: #A9826E;
+    margin: 0;
+    font-weight: 400;
 }
 
 #cooking {
-  color: rgb(211,35,35);
-  margin: 0;
+    color: rgb(211, 35, 35);
+    margin: 0;
 }
 
 #error-page {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
 
-  margin: 20vw 0;
-  overflow: auto;
+    margin: 20vw 0;
+    overflow: auto;
 }
 
 #error-text {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
 }
