+++ react-app/src/components/Navbar/NavBar.css
@@ -1,5 +1,5 @@
 .gradient {
-    background: linear-gradient(180deg,#000000b3,#0000004d 72.92%,#0000);
+    background: linear-gradient(180deg, #000000b3, #0000004d 72.92%, #0000);
     width: 100%;
     top: 0;
     z-index: 10;
@@ -203,15 +203,15 @@
     padding: 8px;
     line-height: 20px;
     border-radius: 4px;
-    font-family: 'Open Sans','Helvetica Neue',Helvetica,Arial,sans-serif;
+    font-family: 'Open Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
     font-size: 14px;
     font-weight: 600;
     letter-spacing: 0px;
 }
 
 .header-right-button:hover {
-    background-color: rgba(255, 255,255, 0.3);
-  }
+    background-color: rgba(255, 255, 255, 0.3);
+}
 
 .top-button-wht {
     color: white;
@@ -328,17 +328,18 @@
 }
 
 .login-button-blk {
-    border: 1px solid rgba(200,201,202,1);
+    border: 1px solid rgba(200, 201, 202, 1);
     /* width: 78px; */
     background-color: transparent;
     transition: 0.8s ease background-color;
 }
 
 .login-button:hover {
-    background-color: rgba(255, 255,255, 0.2);
+    background-color: rgba(255, 255, 255, 0.2);
 }
+
 .login-button-blk:hover {
-    background-color: rgba(255, 255,255, 0.2);
+    background-color: rgba(255, 255, 255, 0.2);
 }
 
 .signup-button:hover {
@@ -348,7 +349,7 @@
 
 .signup-button {
     border: none;
-    background-color: rgb(224,7,7);
+    background-color: rgb(224, 7, 7);
     margin: 0;
     margin-left: 12px;
     transition: 0.8s ease background-color;
@@ -357,7 +358,7 @@
 /************************ homeBanner **************************/
 
 .homepage-banner-container {
-    background: linear-gradient(90deg,#0006,#0000001a 55%);
+    background: linear-gradient(90deg, #0006, #0000001a 55%);
     display: flex;
     flex-direction: column;
     position: absolute;
@@ -415,9 +416,9 @@
 .homepage-button {
     padding: 12px 32px;
     border: none;
-    background-color: rgb(224,7,7);
+    background-color: rgb(224, 7, 7);
     color: white;
-    font-family: 'Open Sans','Helvetica Neue',Helvetica,Arial,sans-serif;
+    font-family: 'Open Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
     font-size: 16px;
     font-weight: 500;
     display: flex;
@@ -468,7 +469,7 @@
     color: white;
     font-size: 14px;
     font-weight: 500;
-    font-family: 'Open Sans','Helvetica Neue',Helvetica,Arial,sans-serif;
+    font-family: 'Open Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
     background-color: transparent;
     border: none;
     padding: 16px 16px 20px 16px;
@@ -479,7 +480,7 @@
 }
 
 .header-bottom-buttons:hover {
-    border-bottom: 4px solid rgb(224,7,7);
+    border-bottom: 4px solid rgb(224, 7, 7);
 }
 
 .popular-button-blk,
@@ -550,7 +551,7 @@
     background-color: white;
 } */
 
-.search-location-input:focus + .tooltip-location {
+.search-location-input:focus+.tooltip-location {
     visibility: visible;
     opacity: 1;
     z-index: 40;
@@ -569,7 +570,7 @@
 }
 
 .tooltip-under-search:hover .header-bottom-buttons {
-    border-bottom: 4px solid rgb(224,7,7);
+    border-bottom: 4px solid rgb(224, 7, 7);
 }
 
 .tooltip-icons {
@@ -580,7 +581,7 @@
 .dropdown-display-text {
     font-size: 14px;
     font-weight: 600;
-    font-family: 'Open Sans','Helvetica Neue',Helvetica,Arial,sans-serif;
+    font-family: 'Open Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
     display: flex;
     align-items: center;
     margin: 0 5px;
