+++ react-app/src/components/User/UserPage.css
@@ -1,7 +1,3 @@
-/* * {
-  font-family: 'Helvetica Neue', 'Helvetica', Arial, sans-serif;
-} */
-
 #bar-img {
     margin-right: 10px;
 }
