+++ react-app/src/components/User/UserReviews.css
@@ -144,6 +144,7 @@
     opacity: 1;
     cursor: pointer;
 }
+
 .user-review-img-container:hover {
     opacity: 0.6;
 }
