+++ react-app/src/context/Modal.css
@@ -1,29 +1,29 @@
 #modal {
-  position: fixed;
-  top: 0;
-  right: 20vw;
-  left: 20vw;
-  bottom: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
+    position: fixed;
+    top: 0;
+    right: 20vw;
+    left: 20vw;
+    bottom: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
 }
 
 #modal-background {
-  position: fixed;
-  top: 0;
-  right: 0;
-  left: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.7);
+    position: fixed;
+    top: 0;
+    right: 0;
+    left: 0;
+    bottom: 0;
+    background-color: rgba(0, 0, 0, 0.7);
 }
 
 #modal-content {
-  position: absolute;
-  background-color:white;
-  border-radius: 12px;
-  width: 100%;
-  max-width: 920px;
-  min-width: 700px;
-  padding: 20px;
+    position: absolute;
+    background-color: white;
+    border-radius: 12px;
+    width: 100%;
+    max-width: 920px;
+    min-width: 700px;
+    padding: 20px;
 }
