+++ react-app/src/context/Modal.js
@@ -5,34 +5,34 @@ import './Modal.css';
 const ModalContext = React.createContext();
 
 export function ModalProvider({ children }) {
-  const modalRef = useRef();
-  const [value, setValue] = useState();
+    const modalRef = useRef();
+    const [value, setValue] = useState();
 
-  useEffect(() => {
-    setValue(modalRef.current);
-  }, [])
+    useEffect(() => {
+        setValue(modalRef.current);
+    }, [])
 
-  return (
-    <>
-      <ModalContext.Provider value={value}>
-        {children}
-      </ModalContext.Provider>
-      <div ref={modalRef} />
-    </>
-  );
+    return (
+        <>
+            <ModalContext.Provider value={value}>
+                {children}
+            </ModalContext.Provider>
+            <div ref={modalRef} />
+        </>
+    );
 }
 
 export function Modal({ onClose, children }) {
-  const modalNode = useContext(ModalContext);
-  if (!modalNode) return null;
+    const modalNode = useContext(ModalContext);
+    if (!modalNode) return null;
 
-  return ReactDOM.createPortal(
-    <div id="modal">
-      <div id="modal-background" onClick={onClose} />
-      <div id="modal-content">
-        {children}
-      </div>
-    </div>,
-    modalNode
-  );
+    return ReactDOM.createPortal(
+        <div id="modal">
+            <div id="modal-background" onClick={onClose} />
+            <div id="modal-content">
+                {children}
+            </div>
+        </div>,
+        modalNode
+    );
 }
