+++ react-app/src/index.css
@@ -23,6 +23,7 @@
 a {
     text-decoration: none;
 }
+
 a:hover {
     text-decoration: underline;
     cursor: pointer;
