+++ react-app/src/index.js
@@ -10,12 +10,12 @@ import { ModalProvider } from './context/Modal';
 const store = configureStore();
 
 ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-        <ModalProvider>
-          <App />
-        </ModalProvider>
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
+    <React.StrictMode>
+        <Provider store={store}>
+            <ModalProvider>
+                <App />
+            </ModalProvider>
+        </Provider>
+    </React.StrictMode>,
+    document.getElementById('root')
 );
